data_transformer/requirements.txt
datetime
pandas
sqlalchemy
python-dotenv==0.13.0
PyMySQL==0.9.3
urllib3==1.25.8
mysql-connector-python
cryptography
psycopg2-binary
pymongo
pyspark
